已修复 CI 日志中的全部失败点： 恢复 API Key 列表的最近使用 IP 补全。 恢复 request_body_bytes 在 usage log 所有写入、查询和扫描路径中的 54 列映射，并补回回归测试。 修正趋势平均值的 float64 断言。 契约测试改为确认普通用户创建 API Key 返回 403，保留管理员独占管理权限。 修复重新生成 API Key 在缺省配置下的空指针。

+++ backend/internal/repository/usage_log_repo_breakdown_test.go
@@ -97,6 +97,6 @@ func TestGetUserRequestBodyTrend_SelectsTopUsersBeforeTimeBuckets(t *testing.T) 
 	require.NoError(t, err)
 	require.Len(t, rows, 1)
 	require.Equal(t, int64(7), rows[0].UserID)
-	require.Equal(t, int64(300), rows[0].AvgRequestBodyBytes)
+	require.Equal(t, float64(300), rows[0].AvgRequestBodyBytes)
 	require.NoError(t, mock.ExpectationsWereMet())
 }
